+++ waspc/e2e-test/test-outputs/waspComplexTest-golden/waspComplexTest/.wasp/out/sdk/wasp/client/operations/actions/index.ts
@@ -1,8 +1,8 @@
 import { type ActionFor, createAction } from './core'
-import { MySpecialAction } from 'wasp/server/operations/actions'
+import { MySpecialAction_ext } from 'wasp/server/operations/actions'
 
 // PUBLIC API
-export const mySpecialAction: ActionFor<MySpecialAction> = createAction<MySpecialAction>(
+export const mySpecialAction: ActionFor<MySpecialAction_ext> = createAction<MySpecialAction_ext>(
   'operations/my-special-action',
   ['User'],
 )